data/crawler/requirements.txt
# Core dependencies
scrapy==2.8.0
beautifulsoup4==4.12.0
requests==2.31.0
pandas==2.0.0
pymongo==4.5.0
pyyaml==6.0.1

# Vietnamese language processing
nltk==3.8.1
pyvi==0.1.1
underthesea==6.0.0  # Vietnamese NLP toolkit

# Vector database
pymilvus==2.2.8
sentence-transformers==2.2.2
torch==2.0.1

# Utilities
python-dateutil==2.8.2
python-decouple==3.8
apscheduler==3.10.0
loguru==0.7.0
tqdm==4.65.0

# Proxy and rate limiting
scrapy-rotating-proxies==0.6.2
scrapy-deltafetch==2.0.1
scrapy-fake-useragent==1.4.4

# Testing
pytest==7.3.1
pytest-mock==3.10.0
responses==0.23.1
pytest-cov==4.1.0 